PRSD-835: Moves summary-related view models into summaryModels
