+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/forms/pages/ConfirmIdentityPage.kt
@@ -5,7 +5,7 @@ import org.springframework.validation.Validator
 import uk.gov.communities.prsdb.webapp.forms.journeys.JourneyData
 import uk.gov.communities.prsdb.webapp.helpers.LandlordRegistrationJourneyDataHelper
 import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.FormModel
-import uk.gov.communities.prsdb.webapp.models.viewModels.SummaryListRowViewModel
+import uk.gov.communities.prsdb.webapp.models.viewModels.summaryModels.SummaryListRowViewModel
 import kotlin.reflect.KClass
 
 class ConfirmIdentityPage(
