+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/summaryModels/SummaryListRowViewModel.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.viewModels
+package uk.gov.communities.prsdb.webapp.models.viewModels.summaryModels
 
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.toJavaLocalDate
