+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/summaryModels/RegisteredPropertyViewModelTests.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.viewModels
+package uk.gov.communities.prsdb.webapp.models.viewModels.summaryModels
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
